+++ api/src/test/scala/dx/api/InstanceTypeDBTest.scala
@@ -2,11 +2,10 @@ package dx.api
 
 import Tags.{ApiTest, EdgeTest}
 import dx.api.InstanceTypeDB.instanceTypeDBFormat
-import dx.util.JsUtils
+import dx.util.{JsUtils, Logger}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 import spray.json._
-import dx.util.Logger
 
 class InstanceTypeDBTest extends AnyFlatSpec with Matchers {
 
@@ -166,12 +165,6 @@ class InstanceTypeDBTest extends AnyFlatSpec with Matchers {
     }
   }
 
-  it should "upgrade to v2 when specifying system requirements with RAM" in {
-    testDb.selectOptimal(InstanceTypeRequest(minMemoryMB = Some(29900))) should matchPattern {
-      case Some(instanceType: DxInstanceType) if instanceType.name == "mem1_ssd1_v2_x16" =>
-    }
-  }
-
   it should "keep v1 instance because v2 is not available in the DB" in {
     testDb
       .selectOptimal(InstanceTypeRequest(minCpu = Some(2), minMemoryMB = Some(7000))) should matchPattern {
@@ -180,7 +173,17 @@ class InstanceTypeDBTest extends AnyFlatSpec with Matchers {
   }
 
   it should "issue a warning if requested a v1 instance by ID but v2 is available" in {
-    throw new Exception("This fails because not implemented")
+    testDb.selectByName("mem1_ssd1_x16") should (matchPattern {
+      case Some(instanceType: DxInstanceType) if instanceType.name == "mem1_ssd1_x16" =>
+    } and equal {
+      Logger.get.warning(
+          s"""
+             |WARNING: an older version of the instance mem1_ssd1_x16 is specified.
+             |Please consider upgrading to a mem1_ssd1_v2_x16 instance
+             |""".stripMargin
+      )
+    })
+
   }
 
   it should "respect requests for GPU instances" taggedAs EdgeTest in {